parser-agent/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiokafka>=0.8.1
pdfplumber>=0.10.2
openai>=1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
